:lipstick: fix: border pool details fixed
