+++ views/pool-details/additional-info/components/accordion/accordion.type.ts
