+++ views/pool-details/additional-info/components/detail-tabs/pool-details-tabs-item.tsx
