+++ views/pool-details/additional-info/components/detail-tabs/pool-details-tabs.types.ts
